backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Audio Processing
openai-whisper==20231117
librosa==0.10.0
pydub==0.25.1
ffmpeg-python==0.2.1

# NLP & Summarization
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
nltk==3.8.1
spacy==3.7.2

# Document Export
python-docx==0.8.11
fpdf2==2.7.0
markdown==3.5.1

# Async Tasks (Optional)
celery==5.3.4
redis==5.0.1

# Database (Optional)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# API & HTTP
requests==2.31.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging & Monitoring
python-json-logger==2.0.7

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
